Rename maps activity to editLocation, refactor maps to use MVP
