+++ app/src/main/java/org/wit/hillfort/views/editLocation/EditLocationView.kt
@@ -1,0 +1,49 @@
+package org.wit.hillfort.views.editLocation
+
+import android.os.Bundle
+import android.support.v7.app.AppCompatActivity
+import com.google.android.gms.maps.GoogleMap
+import com.google.android.gms.maps.SupportMapFragment
+import com.google.android.gms.maps.model.Marker
+import org.wit.hillfort.R
+
+class EditLocationView : AppCompatActivity(), GoogleMap.OnMarkerDragListener, GoogleMap.OnMarkerClickListener {
+  private lateinit var map: GoogleMap
+  private lateinit var presenter: EditLocationPresenter
+
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(savedInstanceState)
+    setContentView(R.layout.activity_maps)
+
+    presenter = EditLocationPresenter(this)
+
+    // Obtain the SupportMapFragment and get notified when the map is ready to be used.
+    val mapFragment = supportFragmentManager
+        .findFragmentById(R.id.map) as SupportMapFragment
+    mapFragment.getMapAsync {
+      map = it
+      map.setOnMarkerDragListener(this)
+      map.setOnMarkerClickListener(this)
+      presenter.initMap(map)
+    }
+  }
+
+  override fun onMarkerDragStart(marker: Marker) {
+  }
+
+  override fun onMarkerDrag(marker: Marker) {
+  }
+
+  override fun onMarkerDragEnd(marker: Marker) {
+    presenter.doUpdateLocation(marker.position.latitude, marker.position.longitude, map.cameraPosition.zoom)
+  }
+
+  override fun onMarkerClick(marker: Marker): Boolean {
+    presenter.doUpdateMarker(marker)
+    return false
+  }
+
+  override fun onBackPressed() {
+    presenter.doOnBackPressed()
+  }
+}
